pref:蓝盾国际化方案优化 #8211

+++ src/backend/ci/core/store/biz-store/src/main/kotlin/com/tencent/devops/store/service/common/impl/CategoryServiceImpl.kt
@@ -192,7 +192,7 @@ class CategoryServiceImpl @Autowired constructor(
     override fun addCategoryToCategoryList(it: Record, categoryList: MutableList<Category>) {
         val categoryCode = it[KEY_CATEGORY_CODE] as String
         val categoryName = it[KEY_CATEGORY_NAME] as String
-        val type = it[KEY_CATEGORY_TYPE] as Int
+        val type = (it[KEY_CATEGORY_TYPE] as Byte).toInt()
         val categoryLanName = I18nUtil.getCodeLanMessage(
             messageCode = "${StoreTypeEnum.getStoreType(type)}.category.$categoryCode",
             defaultMessage = categoryName
